Add stopAllWorkers to devtools protocol
This CL adds ServiceWorker.stopAllWorkers() to devtools protocol for
stopping all service workers before running perf benchmark for PWA.

Design doc of perf benchmark for PWA:
https://docs.google.com/document/d/1Nf97CVp1X7aSqvAspyJ7yOCDyr1osUNrnfrGwZ_Yuuo/edit?usp=sharing


Bug: 736697

+++ ipc/ipc_mojo_message_helper.cc
@@ -33,7 +33,7 @@ bool MojoMessageHelper::ReadMessagePipeFrom(
   MessageAttachment::Type type =
       static_cast<MessageAttachment*>(attachment.get())->GetType();
   if (type != MessageAttachment::Type::MOJO_HANDLE) {
-    LOG(ERROR) << "Unxpected attachment type:" << static_cast<int>(type);
+    LOG(ERROR) << "Unxpected attachment type:" << type;
     return false;
   }
